Made changes to improve startup experience.
Log files are now no longer created wherever the program is called from,
but are instead created in a folder set by the user. This prevents the
log files from being created at shell:startup, if the bat file is placed
there.

Added lines to the README.md to accomodate changes that have taken
place in the .env file.

Added a more detailed description explaining how to set up the file to
run at startup.